# EMSserver/ems/src/main/java/com/obu/ems/dto/RegisterRequest.java
package com.obu.ems.dto;

import jakarta.validation.constraints.*;
import lombok.Data;

@Data
public class RegisterRequest {
    @NotBlank @Size(min = 3, max = 100)
    private String username;

    @NotBlank @Size(min = 6)
    private String password;

    @NotBlank
    private String firstName;

    @NotBlank
    private String lastName;

    @NotBlank
    private String studentNumber;

    @NotNull
    private Long degreeId;
}