# EMSserver/ems/src/main/java/com/obu/ems/dto/UpdateEnrollmentRequest.java
package com.obu.ems.dto;

import com.obu.ems.model.Enrollment;
import jakarta.validation.constraints.NotNull;
import lombok.Data;

@Data
public class UpdateEnrollmentRequest {
    @NotNull
    private Enrollment.Status status;
}